Revert "Rename flann directory and module to cvflann."
This reverts commit e84a089fb273416878461c328f8bb90d9d7389c0.

It looks like the real problem was the use of a FindOpenCV2.cmake
module that was adding /usr/local/include/opencv2/ to the include_dirs
path. Instead, only /usr/local/include should be in the path, and
any #include lines for opencv should explicitly use the "opencv2"
prefix.

That is, rather than using "#include <opencv.hpp>", one should
use "#include <opencv2/opencv.hpp>".

Thanks to Marius on the PCL-users list for the pointer to
the root of the problem.
