+++ apps/haartraining/CMakeLists.txt
@@ -2,7 +2,7 @@ if(IOS OR ANDROID)
   return()
 endif()
 
-SET(OPENCV_HAARTRAINING_DEPS opencv_core opencv_imgproc opencv_highgui opencv_objdetect opencv_calib3d opencv_video opencv_features2d opencv_cvflann opencv_legacy)
+SET(OPENCV_HAARTRAINING_DEPS opencv_core opencv_imgproc opencv_highgui opencv_objdetect opencv_calib3d opencv_video opencv_features2d opencv_flann opencv_legacy)
 ocv_check_dependencies(${OPENCV_HAARTRAINING_DEPS})
 
 if(NOT OCV_DEPENDENCIES_FOUND)
