+++ doc/CMakeLists.txt
@@ -11,7 +11,7 @@ if(BUILD_DOCS AND HAVE_SPHINX)
 
   project(opencv_docs)
 
-  set(OPENCV2_BASE_MODULES core imgproc highgui video calib3d features2d objdetect ml cvflann gpu photo stitching nonfree contrib legacy)
+  set(OPENCV2_BASE_MODULES core imgproc highgui video calib3d features2d objdetect ml flann gpu photo stitching nonfree contrib legacy)
 
   # build lists of modules to be documented
   set(OPENCV2_MODULES "")
