+++ include/opencv/cv.h
@@ -66,7 +66,7 @@
 #include "opencv2/imgproc/imgproc.hpp"
 #include "opencv2/video/tracking.hpp"
 #include "opencv2/features2d/features2d.hpp"
-#include "opencv2/cvflann/flann.hpp"
+#include "opencv2/flann/flann.hpp"
 #include "opencv2/calib3d/calib3d.hpp"
 #include "opencv2/objdetect/objdetect.hpp"
 #include "opencv2/legacy/compat.hpp"
