+++ include/opencv2/opencv.hpp
@@ -45,7 +45,7 @@
 
 #include "opencv2/core/core_c.h"
 #include "opencv2/core/core.hpp"
-#include "opencv2/cvflann/miniflann.hpp"
+#include "opencv2/flann/miniflann.hpp"
 #include "opencv2/imgproc/imgproc_c.h"
 #include "opencv2/imgproc/imgproc.hpp"
 #include "opencv2/photo/photo.hpp"
