+++ modules/features2d/CMakeLists.txt
@@ -1,3 +1,3 @@
 set(the_description "2D Features Framework")
-ocv_define_module(features2d opencv_imgproc opencv_cvflann OPTIONAL opencv_highgui)
+ocv_define_module(features2d opencv_imgproc opencv_flann OPTIONAL opencv_highgui)
