+++ modules/flann/CMakeLists.txt
@@ -1,3 +1,3 @@
 set(the_description "Clustering and Search in Multi-Dimensional Spaces")
-ocv_define_module(cvflann opencv_core)
+ocv_define_module(flann opencv_core)
