+++ modules/flann/include/opencv2/flann/miniflann.hpp
@@ -46,7 +46,7 @@
 #ifdef __cplusplus
 
 #include "opencv2/core/core.hpp"
-#include "opencv2/cvflann/defines.h"
+#include "opencv2/flann/defines.h"
 
 namespace cv
 {
