+++ modules/flann/src/flann.cpp
@@ -27,7 +27,7 @@
  *************************************************************************/
 
 #include "precomp.hpp"
-#include "opencv2/cvflann/flann.hpp"
+#include "opencv2/flann/flann.hpp"
 
 namespace cvflann
 {
