+++ modules/java/generator/gen_javadoc.py
@@ -1,5 +1,5 @@
 import os, sys, re, string, glob
-allmodules = ["core", "cvflann", "imgproc", "ml", "highgui", "video", "features2d", "calib3d", "objdetect", "legacy", "contrib", "gpu", "androidcamera", "haartraining", "java", "python", "stitching", "traincascade", "ts", "photo", "videostab"]
+allmodules = ["core", "flann", "imgproc", "ml", "highgui", "video", "features2d", "calib3d", "objdetect", "legacy", "contrib", "gpu", "androidcamera", "haartraining", "java", "python", "stitching", "traincascade", "ts", "photo", "videostab"]
 verbose = False
 show_warnings = True
 show_errors = True
