+++ modules/java/generator/rst_parser.py
@@ -1,5 +1,5 @@
 import os, sys, re, string, fnmatch
-allmodules = ["core", "cvflann", "imgproc", "ml", "highgui", "video", "features2d", "calib3d", "objdetect", "legacy", "contrib", "gpu", "androidcamera", "java", "python", "stitching", "ts", "photo", "nonfree", "videostab"]
+allmodules = ["core", "flann", "imgproc", "ml", "highgui", "video", "features2d", "calib3d", "objdetect", "legacy", "contrib", "gpu", "androidcamera", "java", "python", "stitching", "ts", "photo", "nonfree", "videostab"]
 verbose = False
 show_warnings = True
 show_errors = True
