+++ modules/python/src2/cv2.cpp
@@ -10,7 +10,7 @@
 
 #include "opencv2/core/core.hpp"
 #include "opencv2/contrib/contrib.hpp"
-#include "opencv2/cvflann/miniflann.hpp"
+#include "opencv2/flann/miniflann.hpp"
 #include "opencv2/imgproc/imgproc.hpp"
 #include "opencv2/calib3d/calib3d.hpp"
 #include "opencv2/ml/ml.hpp"
