+++ samples/cpp/image.cpp
@@ -2,7 +2,7 @@
 #include <iostream>
 #include "opencv2/imgproc/imgproc.hpp"
 #include "opencv2/highgui/highgui.hpp"
-#include "opencv2/cvflann/miniflann.hpp"
+#include "opencv2/flann/miniflann.hpp"
 
 using namespace cv; // all the new API is put into "cv" namespace. Export its content
 using namespace std;
